chore: export vsop earth data
some bundlers do not yet recognize the node package.json exports but
some of them do. In order to regain compatibility with those who don't
or won't support this (e.g. browserify) there is a need to pass on the
commonjs file for e.g. date-chinese such that node and browserify find
the module, node via package exports and browserify via file path.