saturnmoons, jupitermoons - add Z-coord output
Add the `Z`-coordinate to the computed positions of the moons, so that
transits, eclipses, occultations, and shadow transits can be more easily
computed.

At this moment we do not know of any published "official" Z-values to
check against, so we use whatever values would satisfy the test at the
moment and hard-code them into the test files.